pyproject.toml
[project]
name = "aur2"
version = "0.1.0"
description = "Agentic scaffolding for knowledge work"
requires-python = ">=3.12"
dependencies = [
    "click>=8.0",
    "python-dotenv>=1.0",
]

[project.scripts]
aur2 = "aur2.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aur2"]
